Did some final organization, having issues with debug symbols
For some reason building with debug flags doesn't produce debug symbols with clang...

Did more cleanup in the project, I got the organization setup pretty much finalized so keeping like this from now on.
Also sticking with STL not using zpl as there are issues with missing symbols even with proper include...
Could uze zlib but just keeping things simple, since performance is not super high priority just getting mvp PAL setup for C was.

+++ Interpreter/project/Lexer/ProtoLexer.h
@@ -6,7 +6,7 @@
 
 struct Lexer
 {
-	u32   Index;
+	u64   Index;
 	str   Contents;	
 	uDM   ContentSize;
 	schar CurrentChar;
